Separate handler methods from context API
