+++ .changelog/unreleased/breaking-changes/539-separate-handler-functions.md
@@ -1,0 +1,2 @@
+- Separate validation/execution handlers from context API
+  ([#539](https://github.com/cosmos/ibc-rs/issues/539))
