+++ crates/ibc/src/core/handler.rs
@@ -1,36 +1,167 @@
-use ibc_proto::google::protobuf::Any;
-
+use super::ics02_client::handler::{create_client, update_client, upgrade_client};
+use super::ics02_client::msgs::{ClientMsg, MsgUpdateOrMisbehaviour};
+use super::ics03_connection::handler::{
+    conn_open_ack, conn_open_confirm, conn_open_init, conn_open_try,
+};
+use super::ics03_connection::msgs::ConnectionMsg;
+use super::ics04_channel::error::ChannelError;
+use super::ics04_channel::handler::acknowledgement::{
+    acknowledgement_packet_execute, acknowledgement_packet_validate,
+};
+use super::ics04_channel::handler::chan_close_confirm::{
+    chan_close_confirm_execute, chan_close_confirm_validate,
+};
+use super::ics04_channel::handler::chan_close_init::{
+    chan_close_init_execute, chan_close_init_validate,
+};
+use super::ics04_channel::handler::chan_open_ack::{chan_open_ack_execute, chan_open_ack_validate};
+use super::ics04_channel::handler::chan_open_confirm::{
+    chan_open_confirm_execute, chan_open_confirm_validate,
+};
+use super::ics04_channel::handler::chan_open_init::{
+    chan_open_init_execute, chan_open_init_validate,
+};
+use super::ics04_channel::handler::chan_open_try::{chan_open_try_execute, chan_open_try_validate};
+use super::ics04_channel::handler::recv_packet::{recv_packet_execute, recv_packet_validate};
+use super::ics04_channel::handler::timeout::{
+    timeout_packet_execute, timeout_packet_validate, TimeoutMsgType,
+};
+use super::ics04_channel::msgs::{ChannelMsg, PacketMsg};
+use super::ContextError;
 use super::{msgs::MsgEnvelope, ExecutionContext, RouterError, ValidationContext};
 
-/// Entrypoint which only performs message validation
-///
-/// If a transaction contains `n` messages `m_1` ... `m_n`, then
-/// they MUST be processed as follows:
-///     validate(m_1), execute(m_1), ..., validate(m_n), execute(m_n)
-/// That is, the state transition of message `i` must be applied before
-/// message `i+1` is validated. This is equivalent to calling
-/// `dispatch()` on each successively.
-pub fn validate<Ctx>(ctx: &Ctx, message: Any) -> Result<(), RouterError>
+/// Validation handler that dispatches messages to the correct core ICS handler.
+pub(crate) fn validation_handler<Ctx>(ctx: &Ctx, msg: MsgEnvelope) -> Result<(), RouterError>
 where
     Ctx: ValidationContext,
 {
-    let envelope: MsgEnvelope = message.try_into()?;
-    ctx.validate(envelope)
+    match msg {
+        MsgEnvelope::Client(msg) => match msg {
+            ClientMsg::CreateClient(msg) => create_client::validate(ctx, msg),
+            ClientMsg::UpdateClient(msg) => {
+                update_client::validate(ctx, MsgUpdateOrMisbehaviour::UpdateClient(msg))
+            }
+            ClientMsg::Misbehaviour(msg) => {
+                update_client::validate(ctx, MsgUpdateOrMisbehaviour::Misbehaviour(msg))
+            }
+            ClientMsg::UpgradeClient(msg) => upgrade_client::validate(ctx, msg),
+        }
+        .map_err(RouterError::ContextError),
+        MsgEnvelope::Connection(msg) => match msg {
+            ConnectionMsg::OpenInit(msg) => conn_open_init::validate(ctx, msg),
+            ConnectionMsg::OpenTry(msg) => conn_open_try::validate(ctx, msg),
+            ConnectionMsg::OpenAck(msg) => conn_open_ack::validate(ctx, msg),
+            ConnectionMsg::OpenConfirm(ref msg) => conn_open_confirm::validate(ctx, msg),
+        }
+        .map_err(RouterError::ContextError),
+        MsgEnvelope::Channel(msg) => {
+            let module_id = ctx
+                .lookup_module_channel(&msg)
+                .map_err(ContextError::from)?;
+            if !ctx.has_route(&module_id) {
+                return Err(ChannelError::RouteNotFound)
+                    .map_err(ContextError::ChannelError)
+                    .map_err(RouterError::ContextError);
+            }
+
+            match msg {
+                ChannelMsg::OpenInit(msg) => chan_open_init_validate(ctx, module_id, msg),
+                ChannelMsg::OpenTry(msg) => chan_open_try_validate(ctx, module_id, msg),
+                ChannelMsg::OpenAck(msg) => chan_open_ack_validate(ctx, module_id, msg),
+                ChannelMsg::OpenConfirm(msg) => chan_open_confirm_validate(ctx, module_id, msg),
+                ChannelMsg::CloseInit(msg) => chan_close_init_validate(ctx, module_id, msg),
+                ChannelMsg::CloseConfirm(msg) => chan_close_confirm_validate(ctx, module_id, msg),
+            }
+            .map_err(RouterError::ContextError)
+        }
+        MsgEnvelope::Packet(msg) => {
+            let module_id = ctx.lookup_module_packet(&msg).map_err(ContextError::from)?;
+            if !ctx.has_route(&module_id) {
+                return Err(ChannelError::RouteNotFound)
+                    .map_err(ContextError::ChannelError)
+                    .map_err(RouterError::ContextError);
+            }
+
+            match msg {
+                PacketMsg::Recv(msg) => recv_packet_validate(ctx, msg),
+                PacketMsg::Ack(msg) => acknowledgement_packet_validate(ctx, module_id, msg),
+                PacketMsg::Timeout(msg) => {
+                    timeout_packet_validate(ctx, module_id, TimeoutMsgType::Timeout(msg))
+                }
+                PacketMsg::TimeoutOnClose(msg) => {
+                    timeout_packet_validate(ctx, module_id, TimeoutMsgType::TimeoutOnClose(msg))
+                }
+            }
+            .map_err(RouterError::ContextError)
+        }
+    }
 }
 
-/// Entrypoint which only performs message execution
-pub fn execute<Ctx>(ctx: &mut Ctx, message: Any) -> Result<(), RouterError>
+/// Execution handler that dispatches messages to the correct core ICS handler.
+pub(crate) fn execution_handler<Ctx>(ctx: &mut Ctx, msg: MsgEnvelope) -> Result<(), RouterError>
 where
     Ctx: ExecutionContext,
 {
-    let envelope: MsgEnvelope = message.try_into()?;
-    ctx.execute(envelope)
-}
+    match msg {
+        MsgEnvelope::Client(msg) => match msg {
+            ClientMsg::CreateClient(msg) => create_client::execute(ctx, msg),
+            ClientMsg::UpdateClient(msg) => {
+                update_client::execute(ctx, MsgUpdateOrMisbehaviour::UpdateClient(msg))
+            }
+            ClientMsg::Misbehaviour(msg) => {
+                update_client::execute(ctx, MsgUpdateOrMisbehaviour::Misbehaviour(msg))
+            }
+            ClientMsg::UpgradeClient(msg) => upgrade_client::execute(ctx, msg),
+        }
+        .map_err(RouterError::ContextError),
+        MsgEnvelope::Connection(msg) => match msg {
+            ConnectionMsg::OpenInit(msg) => conn_open_init::execute(ctx, msg),
+            ConnectionMsg::OpenTry(msg) => conn_open_try::execute(ctx, msg),
+            ConnectionMsg::OpenAck(msg) => conn_open_ack::execute(ctx, msg),
+            ConnectionMsg::OpenConfirm(ref msg) => conn_open_confirm::execute(ctx, msg),
+        }
+        .map_err(RouterError::ContextError),
+        MsgEnvelope::Channel(msg) => {
+            let module_id = ctx
+                .lookup_module_channel(&msg)
+                .map_err(ContextError::from)?;
+            if !ctx.has_route(&module_id) {
+                return Err(ChannelError::RouteNotFound)
+                    .map_err(ContextError::ChannelError)
+                    .map_err(RouterError::ContextError);
+            }
 
-/// Entrypoint which performs both validation and message execution
-pub fn dispatch(ctx: &mut impl ExecutionContext, msg: MsgEnvelope) -> Result<(), RouterError> {
-    ctx.validate(msg.clone())?;
-    ctx.execute(msg)
+            match msg {
+                ChannelMsg::OpenInit(msg) => chan_open_init_execute(ctx, module_id, msg),
+                ChannelMsg::OpenTry(msg) => chan_open_try_execute(ctx, module_id, msg),
+                ChannelMsg::OpenAck(msg) => chan_open_ack_execute(ctx, module_id, msg),
+                ChannelMsg::OpenConfirm(msg) => chan_open_confirm_execute(ctx, module_id, msg),
+                ChannelMsg::CloseInit(msg) => chan_close_init_execute(ctx, module_id, msg),
+                ChannelMsg::CloseConfirm(msg) => chan_close_confirm_execute(ctx, module_id, msg),
+            }
+            .map_err(RouterError::ContextError)
+        }
+        MsgEnvelope::Packet(msg) => {
+            let module_id = ctx.lookup_module_packet(&msg).map_err(ContextError::from)?;
+            if !ctx.has_route(&module_id) {
+                return Err(ChannelError::RouteNotFound)
+                    .map_err(ContextError::ChannelError)
+                    .map_err(RouterError::ContextError);
+            }
+
+            match msg {
+                PacketMsg::Recv(msg) => recv_packet_execute(ctx, module_id, msg),
+                PacketMsg::Ack(msg) => acknowledgement_packet_execute(ctx, module_id, msg),
+                PacketMsg::Timeout(msg) => {
+                    timeout_packet_execute(ctx, module_id, TimeoutMsgType::Timeout(msg))
+                }
+                PacketMsg::TimeoutOnClose(msg) => {
+                    timeout_packet_execute(ctx, module_id, TimeoutMsgType::TimeoutOnClose(msg))
+                }
+            }
+            .map_err(RouterError::ContextError)
+        }
+    }
 }
 
 #[cfg(test)]
