+++ crates/ibc/src/core/ics04_channel/handler/acknowledgement.rs
@@ -1,3 +1,6 @@
+use crate::prelude::*;
+
+use crate::core::events::MessageEvent;
 use crate::core::ics03_connection::connection::State as ConnectionState;
 use crate::core::ics03_connection::delay::verify_conn_delay_passed;
 use crate::core::ics04_channel::channel::{Counterparty, Order, State as ChannelState};
@@ -9,9 +12,109 @@ use crate::core::ics24_host::path::Path;
 use crate::core::ics24_host::path::{
     AckPath, ChannelEndPath, ClientConsensusStatePath, CommitmentPath, SeqAckPath,
 };
-use crate::prelude::*;
+use crate::core::{events::IbcEvent, ics04_channel::events::AcknowledgePacket, router::ModuleId};
+use crate::core::{ContextError, ExecutionContext, ValidationContext};
 
-use crate::core::{ContextError, ValidationContext};
+pub(crate) fn acknowledgement_packet_validate<ValCtx>(
+    ctx_a: &ValCtx,
+    module_id: ModuleId,
+    msg: MsgAcknowledgement,
+) -> Result<(), ContextError>
+where
+    ValCtx: ValidationContext,
+{
+    validate(ctx_a, &msg)?;
+
+    let module = ctx_a
+        .get_route(&module_id)
+        .ok_or(ChannelError::RouteNotFound)?;
+
+    module
+        .on_acknowledgement_packet_validate(&msg.packet, &msg.acknowledgement, &msg.signer)
+        .map_err(ContextError::PacketError)
+}
+
+pub(crate) fn acknowledgement_packet_execute<ExecCtx>(
+    ctx_a: &mut ExecCtx,
+    module_id: ModuleId,
+    msg: MsgAcknowledgement,
+) -> Result<(), ContextError>
+where
+    ExecCtx: ExecutionContext,
+{
+    let chan_end_path_on_a =
+        ChannelEndPath::new(&msg.packet.port_id_on_a, &msg.packet.chan_id_on_a);
+    let chan_end_on_a = ctx_a.channel_end(&chan_end_path_on_a)?;
+    let conn_id_on_a = &chan_end_on_a.connection_hops()[0];
+
+    // In all cases, this event is emitted
+    let event = IbcEvent::AcknowledgePacket(AcknowledgePacket::new(
+        msg.packet.clone(),
+        chan_end_on_a.ordering,
+        conn_id_on_a.clone(),
+    ));
+    ctx_a.emit_ibc_event(IbcEvent::Message(MessageEvent::Channel));
+    ctx_a.emit_ibc_event(event);
+
+    let commitment_path_on_a = CommitmentPath::new(
+        &msg.packet.port_id_on_a,
+        &msg.packet.chan_id_on_a,
+        msg.packet.seq_on_a,
+    );
+
+    // check if we're in the NO-OP case
+    if ctx_a.get_packet_commitment(&commitment_path_on_a).is_err() {
+        // This error indicates that the timeout has already been relayed
+        // or there is a misconfigured relayer attempting to prove a timeout
+        // for a packet never sent. Core IBC will treat this error as a no-op in order to
+        // prevent an entire relay transaction from failing and consuming unnecessary fees.
+        return Ok(());
+    };
+
+    let module = ctx_a
+        .get_route_mut(&module_id)
+        .ok_or(ChannelError::RouteNotFound)?;
+
+    let (extras, cb_result) =
+        module.on_acknowledgement_packet_execute(&msg.packet, &msg.acknowledgement, &msg.signer);
+
+    cb_result?;
+
+    // apply state changes
+    {
+        let commitment_path_on_a = CommitmentPath {
+            port_id: msg.packet.port_id_on_a.clone(),
+            channel_id: msg.packet.chan_id_on_a.clone(),
+            sequence: msg.packet.seq_on_a,
+        };
+        ctx_a.delete_packet_commitment(&commitment_path_on_a)?;
+
+        if let Order::Ordered = chan_end_on_a.ordering {
+            // Note: in validation, we verified that `msg.packet.sequence == nextSeqRecv`
+            // (where `nextSeqRecv` is the value in the store)
+            let seq_ack_path_on_a =
+                SeqAckPath::new(&msg.packet.port_id_on_a, &msg.packet.chan_id_on_a);
+            ctx_a.store_next_sequence_ack(&seq_ack_path_on_a, msg.packet.seq_on_a.increment())?;
+        }
+    }
+
+    // emit events and logs
+    {
+        ctx_a.log_message("success: packet acknowledgement".to_string());
+
+        // Note: Acknowledgement event was emitted at the beginning
+
+        for module_event in extras.events {
+            ctx_a.emit_ibc_event(IbcEvent::Module(module_event));
+        }
+
+        for log_message in extras.log {
+            ctx_a.log_message(log_message);
+        }
+    }
+
+    Ok(())
+}
 
 pub fn validate<Ctx>(ctx_a: &Ctx, msg: &MsgAcknowledgement) -> Result<(), ContextError>
 where
@@ -114,13 +217,7 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::core::ics04_channel::commitment::compute_packet_commitment;
-    use crate::core::ics04_channel::handler::acknowledgement::validate;
-    use crate::core::ics24_host::identifier::ChannelId;
-    use crate::core::ics24_host::identifier::PortId;
-    use crate::core::timestamp::Timestamp;
-    use crate::core::ExecutionContext;
-    use crate::prelude::*;
+    use super::*;
     use rstest::*;
     use test_log::test;
 
@@ -134,38 +231,62 @@ mod tests {
     use crate::core::ics04_channel::msgs::acknowledgement::test_util::get_dummy_raw_msg_acknowledgement;
     use crate::core::ics04_channel::msgs::acknowledgement::MsgAcknowledgement;
     use crate::core::ics04_channel::Version;
+    use crate::core::ics24_host::identifier::ChannelId;
+    use crate::core::ics24_host::identifier::PortId;
     use crate::core::ics24_host::identifier::{ClientId, ConnectionId};
+    use crate::core::timestamp::Timestamp;
     use crate::core::timestamp::ZERO_DURATION;
-    use crate::mock::context::MockContext;
 
-    pub struct Fixture {
-        pub context: MockContext,
-        pub client_height: Height,
-        pub msg: MsgAcknowledgement,
-        pub packet_commitment: PacketCommitment,
-        pub conn_end_on_a: ConnectionEnd,
-        pub chan_end_on_a: ChannelEnd,
+    use crate::mock::context::MockContext;
+    use crate::{applications::transfer::MODULE_ID_STR, test_utils::DummyTransferModule};
+
+    struct Fixture {
+        ctx: MockContext,
+        module_id: ModuleId,
+        msg: MsgAcknowledgement,
+        packet_commitment: PacketCommitment,
+        conn_end_on_a: ConnectionEnd,
+        chan_end_on_a_ordered: ChannelEnd,
+        chan_end_on_a_unordered: ChannelEnd,
     }
 
     #[fixture]
     fn fixture() -> Fixture {
-        let context = MockContext::default();
         let client_height = Height::new(0, 2).unwrap();
+        let mut ctx = MockContext::default().with_client(&ClientId::default(), client_height);
+
+        let module_id: ModuleId = ModuleId::new(MODULE_ID_STR.to_string());
+        let module = DummyTransferModule::new();
+        ctx.add_route(module_id.clone(), module).unwrap();
+
         let msg = MsgAcknowledgement::try_from(get_dummy_raw_msg_acknowledgement(
             client_height.revision_height(),
         ))
         .unwrap();
+
         let packet = msg.packet.clone();
 
         let packet_commitment = compute_packet_commitment(
-            &packet.data,
-            &packet.timeout_height_on_b,
-            &packet.timeout_timestamp_on_b,
+            &msg.packet.data,
+            &msg.packet.timeout_height_on_b,
+            &msg.packet.timeout_timestamp_on_b,
         );
 
-        let chan_end_on_a = ChannelEnd::new(
+        let chan_end_on_a_unordered = ChannelEnd::new(
             State::Open,
-            Order::default(),
+            Order::Unordered,
+            Counterparty::new(
+                packet.port_id_on_b.clone(),
+                Some(packet.chan_id_on_b.clone()),
+            ),
+            vec![ConnectionId::default()],
+            Version::new("ics20-1".to_string()),
+        )
+        .unwrap();
+
+        let chan_end_on_a_ordered = ChannelEnd::new(
+            State::Open,
+            Order::Ordered,
             Counterparty::new(packet.port_id_on_b.clone(), Some(packet.chan_id_on_b)),
             vec![ConnectionId::default()],
             Version::new("ics20-1".to_string()),
@@ -186,20 +307,21 @@ mod tests {
         .unwrap();
 
         Fixture {
-            context,
-            client_height,
+            ctx,
+            module_id,
             msg,
             packet_commitment,
             conn_end_on_a,
-            chan_end_on_a,
+            chan_end_on_a_unordered,
+            chan_end_on_a_ordered,
         }
     }
 
     #[rstest]
     fn ack_fail_no_channel(fixture: Fixture) {
-        let Fixture { context, msg, .. } = fixture;
+        let Fixture { ctx, msg, .. } = fixture;
 
-        let res = validate(&context, &msg);
+        let res = validate(&ctx, &msg);
 
         assert!(
             res.is_err(),
@@ -211,19 +333,23 @@ mod tests {
     #[rstest]
     fn ack_success_no_packet_commitment(fixture: Fixture) {
         let Fixture {
-            context,
+            ctx,
             msg,
             conn_end_on_a,
-            chan_end_on_a,
-            client_height,
+            chan_end_on_a_unordered,
             ..
         } = fixture;
-        let context = context
+        let client_height = Height::new(0, 2).unwrap();
+        let ctx = ctx
             .with_client(&ClientId::default(), client_height)
-            .with_channel(PortId::default(), ChannelId::default(), chan_end_on_a)
+            .with_channel(
+                PortId::default(),
+                ChannelId::default(),
+                chan_end_on_a_unordered,
+            )
             .with_connection(ConnectionId::default(), conn_end_on_a);
 
-        let res = validate(&context, &msg);
+        let res = validate(&ctx, &msg);
 
         assert!(
             res.is_ok(),
@@ -234,17 +360,21 @@ mod tests {
     #[rstest]
     fn ack_success_happy_path(fixture: Fixture) {
         let Fixture {
-            context,
+            ctx,
             msg,
             packet_commitment,
             conn_end_on_a,
-            chan_end_on_a,
-            client_height,
+            chan_end_on_a_unordered,
             ..
         } = fixture;
-        let mut context = context
+        let client_height = Height::new(0, 2).unwrap();
+        let mut ctx = ctx
             .with_client(&ClientId::default(), client_height)
-            .with_channel(PortId::default(), ChannelId::default(), chan_end_on_a)
+            .with_channel(
+                PortId::default(),
+                ChannelId::default(),
+                chan_end_on_a_unordered,
+            )
             .with_connection(ConnectionId::default(), conn_end_on_a)
             .with_packet_commitment(
                 msg.packet.port_id_on_a.clone(),
@@ -252,26 +382,98 @@ mod tests {
                 msg.packet.seq_on_a,
                 packet_commitment,
             );
-        context
-            .store_update_time(
-                ClientId::default(),
-                client_height,
-                Timestamp::from_nanoseconds(1000).unwrap(),
-            )
-            .unwrap();
-        context
-            .store_update_height(
-                ClientId::default(),
-                client_height,
-                Height::new(0, 4).unwrap(),
-            )
-            .unwrap();
+        ctx.store_update_time(
+            ClientId::default(),
+            client_height,
+            Timestamp::from_nanoseconds(1000).unwrap(),
+        )
+        .unwrap();
+        ctx.store_update_height(
+            ClientId::default(),
+            client_height,
+            Height::new(0, 4).unwrap(),
+        )
+        .unwrap();
 
-        let res = validate(&context, &msg);
+        let res = validate(&ctx, &msg);
 
         assert!(
             res.is_ok(),
             "Happy path: validation should succeed. err: {res:?}"
         )
+    }
+
+    #[rstest]
+    fn ack_unordered_chan_execute(fixture: Fixture) {
+        let Fixture {
+            ctx,
+            module_id,
+            msg,
+            packet_commitment,
+            conn_end_on_a,
+            chan_end_on_a_unordered,
+            ..
+        } = fixture;
+        let mut ctx = ctx
+            .with_channel(
+                PortId::default(),
+                ChannelId::default(),
+                chan_end_on_a_unordered,
+            )
+            .with_connection(ConnectionId::default(), conn_end_on_a)
+            .with_packet_commitment(
+                msg.packet.port_id_on_a.clone(),
+                msg.packet.chan_id_on_a.clone(),
+                msg.packet.seq_on_a,
+                packet_commitment,
+            );
+
+        let res = acknowledgement_packet_execute(&mut ctx, module_id, msg);
+
+        assert!(res.is_ok());
+
+        assert_eq!(ctx.events.len(), 2);
+        assert!(matches!(
+            ctx.events[0],
+            IbcEvent::Message(MessageEvent::Channel)
+        ));
+        assert!(matches!(ctx.events[1], IbcEvent::AcknowledgePacket(_)));
+    }
+
+    #[rstest]
+    fn ack_ordered_chan_execute(fixture: Fixture) {
+        let Fixture {
+            ctx,
+            module_id,
+            msg,
+            packet_commitment,
+            conn_end_on_a,
+            chan_end_on_a_ordered,
+            ..
+        } = fixture;
+        let mut ctx = ctx
+            .with_channel(
+                PortId::default(),
+                ChannelId::default(),
+                chan_end_on_a_ordered,
+            )
+            .with_connection(ConnectionId::default(), conn_end_on_a)
+            .with_packet_commitment(
+                msg.packet.port_id_on_a.clone(),
+                msg.packet.chan_id_on_a.clone(),
+                msg.packet.seq_on_a,
+                packet_commitment,
+            );
+
+        let res = acknowledgement_packet_execute(&mut ctx, module_id, msg);
+
+        assert!(res.is_ok());
+
+        assert_eq!(ctx.events.len(), 2);
+        assert!(matches!(
+            ctx.events[0],
+            IbcEvent::Message(MessageEvent::Channel)
+        ));
+        assert!(matches!(ctx.events[1], IbcEvent::AcknowledgePacket(_)));
     }
 }
