+++ crates/ibc/src/core/ics04_channel/handler/chan_open_confirm.rs
@@ -1,17 +1,104 @@
 //! Protocol logic specific to ICS4 messages of type `MsgChannelOpenConfirm`.
 
+use crate::prelude::*;
 use ibc_proto::protobuf::Protobuf;
 
+use crate::core::events::{IbcEvent, MessageEvent};
 use crate::core::ics03_connection::connection::State as ConnectionState;
+use crate::core::ics04_channel::channel::State;
 use crate::core::ics04_channel::channel::{ChannelEnd, Counterparty, State as ChannelState};
 use crate::core::ics04_channel::error::ChannelError;
+use crate::core::ics04_channel::events::OpenConfirm;
 use crate::core::ics04_channel::msgs::chan_open_confirm::MsgChannelOpenConfirm;
 use crate::core::ics24_host::path::Path;
 use crate::core::ics24_host::path::{ChannelEndPath, ClientConsensusStatePath};
-use crate::core::{ContextError, ValidationContext};
-use crate::prelude::*;
+use crate::core::router::ModuleId;
+use crate::core::{ContextError, ExecutionContext, ValidationContext};
 
-pub fn validate<Ctx>(ctx_b: &Ctx, msg: &MsgChannelOpenConfirm) -> Result<(), ContextError>
+pub(crate) fn chan_open_confirm_validate<ValCtx>(
+    ctx_b: &ValCtx,
+    module_id: ModuleId,
+    msg: MsgChannelOpenConfirm,
+) -> Result<(), ContextError>
+where
+    ValCtx: ValidationContext,
+{
+    validate(ctx_b, &msg)?;
+
+    let module = ctx_b
+        .get_route(&module_id)
+        .ok_or(ChannelError::RouteNotFound)?;
+    module.on_chan_open_confirm_validate(&msg.port_id_on_b, &msg.chan_id_on_b)?;
+
+    Ok(())
+}
+
+pub(crate) fn chan_open_confirm_execute<ExecCtx>(
+    ctx_b: &mut ExecCtx,
+    module_id: ModuleId,
+    msg: MsgChannelOpenConfirm,
+) -> Result<(), ContextError>
+where
+    ExecCtx: ExecutionContext,
+{
+    let module = ctx_b
+        .get_route_mut(&module_id)
+        .ok_or(ChannelError::RouteNotFound)?;
+
+    let extras = module.on_chan_open_confirm_execute(&msg.port_id_on_b, &msg.chan_id_on_b)?;
+    let chan_end_path_on_b = ChannelEndPath::new(&msg.port_id_on_b, &msg.chan_id_on_b);
+    let chan_end_on_b = ctx_b.channel_end(&chan_end_path_on_b)?;
+
+    // state changes
+    {
+        let chan_end_on_b = {
+            let mut chan_end_on_b = chan_end_on_b.clone();
+            chan_end_on_b.set_state(State::Open);
+
+            chan_end_on_b
+        };
+        ctx_b.store_channel(&chan_end_path_on_b, chan_end_on_b)?;
+    }
+
+    // emit events and logs
+    {
+        ctx_b.log_message("success: channel open confirm".to_string());
+
+        let conn_id_on_b = chan_end_on_b.connection_hops[0].clone();
+        let port_id_on_a = chan_end_on_b.counterparty().port_id.clone();
+        let chan_id_on_a = chan_end_on_b
+            .counterparty()
+            .channel_id
+            .clone()
+            .ok_or(ContextError::ChannelError(ChannelError::Other {
+            description:
+                "internal error: ChannelEnd doesn't have a counterparty channel id in OpenConfirm"
+                    .to_string(),
+        }))?;
+
+        let core_event = IbcEvent::OpenConfirmChannel(OpenConfirm::new(
+            msg.port_id_on_b.clone(),
+            msg.chan_id_on_b.clone(),
+            port_id_on_a,
+            chan_id_on_a,
+            conn_id_on_b,
+        ));
+        ctx_b.emit_ibc_event(IbcEvent::Message(MessageEvent::Channel));
+        ctx_b.emit_ibc_event(core_event);
+
+        for module_event in extras.events {
+            ctx_b.emit_ibc_event(IbcEvent::Module(module_event));
+        }
+
+        for log_message in extras.log {
+            ctx_b.log_message(log_message);
+        }
+    }
+
+    Ok(())
+}
+
+fn validate<Ctx>(ctx_b: &Ctx, msg: &MsgChannelOpenConfirm) -> Result<(), ContextError>
 where
     Ctx: ValidationContext,
 {
@@ -81,9 +168,7 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::core::ics04_channel::handler::chan_open_confirm::validate;
-    use crate::core::ics24_host::identifier::ChannelId;
-    use crate::prelude::*;
+    use super::*;
     use rstest::*;
     use test_log::test;
 
@@ -96,14 +181,21 @@ mod tests {
     use crate::core::ics04_channel::msgs::chan_open_confirm::test_util::get_dummy_raw_msg_chan_open_confirm;
     use crate::core::ics04_channel::msgs::chan_open_confirm::MsgChannelOpenConfirm;
     use crate::core::ics04_channel::Version;
+    use crate::core::ics24_host::identifier::ChannelId;
     use crate::core::ics24_host::identifier::{ClientId, ConnectionId};
     use crate::core::timestamp::ZERO_DURATION;
+    use crate::Height;
+
     use crate::mock::client_state::client_type as mock_client_type;
     use crate::mock::context::MockContext;
-    use crate::Height;
+    use crate::{
+        applications::transfer::MODULE_ID_STR, core::router::ModuleId,
+        test_utils::DummyTransferModule,
+    };
 
     pub struct Fixture {
         pub context: MockContext,
+        pub module_id: ModuleId,
         pub msg: MsgChannelOpenConfirm,
         pub client_id_on_b: ClientId,
         pub conn_id_on_b: ConnectionId,
@@ -115,7 +207,10 @@ mod tests {
     #[fixture]
     fn fixture() -> Fixture {
         let proof_height = 10;
-        let context = MockContext::default();
+        let mut context = MockContext::default();
+        let module = DummyTransferModule::new();
+        let module_id: ModuleId = ModuleId::new(MODULE_ID_STR.to_string());
+        context.add_route(module_id.clone(), module).unwrap();
 
         let client_id_on_b = ClientId::new(mock_client_type(), 45).unwrap();
         let conn_id_on_b = ConnectionId::new(2);
@@ -143,6 +238,7 @@ mod tests {
 
         Fixture {
             context,
+            module_id,
             msg,
             client_id_on_b,
             conn_id_on_b,
@@ -237,5 +333,40 @@ mod tests {
         let res = validate(&context, &msg);
 
         assert!(res.is_ok(), "Validation happy path")
+    }
+
+    #[rstest]
+    fn chan_open_confirm_execute_happy_path(fixture: Fixture) {
+        let Fixture {
+            context,
+            module_id,
+            msg,
+            client_id_on_b,
+            conn_id_on_b,
+            conn_end_on_b,
+            chan_end_on_b,
+            proof_height,
+            ..
+        } = fixture;
+
+        let mut context = context
+            .with_client(&client_id_on_b, Height::new(0, proof_height).unwrap())
+            .with_connection(conn_id_on_b, conn_end_on_b)
+            .with_channel(
+                msg.port_id_on_b.clone(),
+                ChannelId::default(),
+                chan_end_on_b,
+            );
+
+        let res = chan_open_confirm_execute(&mut context, module_id, msg);
+
+        assert!(res.is_ok(), "Execution happy path");
+
+        assert_eq!(context.events.len(), 2);
+        assert!(matches!(
+            context.events[0],
+            IbcEvent::Message(MessageEvent::Channel)
+        ));
+        assert!(matches!(context.events[1], IbcEvent::OpenConfirmChannel(_)));
     }
 }
