+++ crates/ibc/src/core/ics04_channel/handler/timeout.rs
@@ -1,20 +1,151 @@
+use crate::prelude::*;
 use prost::Message;
 
+use crate::core::events::MessageEvent;
 use crate::core::ics03_connection::delay::verify_conn_delay_passed;
 use crate::core::ics04_channel::channel::State;
 use crate::core::ics04_channel::channel::{Counterparty, Order};
 use crate::core::ics04_channel::commitment::compute_packet_commitment;
 use crate::core::ics04_channel::error::ChannelError;
 use crate::core::ics04_channel::error::PacketError;
+use crate::core::ics04_channel::events::ChannelClosed;
 use crate::core::ics04_channel::msgs::timeout::MsgTimeout;
+use crate::core::ics04_channel::msgs::timeout_on_close::MsgTimeoutOnClose;
 use crate::core::ics24_host::path::Path;
 use crate::core::ics24_host::path::{
     ChannelEndPath, ClientConsensusStatePath, CommitmentPath, ReceiptPath, SeqRecvPath,
 };
-use crate::core::{ContextError, ValidationContext};
-use crate::prelude::*;
+use crate::core::{
+    events::IbcEvent,
+    ics04_channel::{events::TimeoutPacket, handler::timeout_on_close},
+    router::ModuleId,
+};
+use crate::core::{ContextError, ExecutionContext, ValidationContext};
 
-pub fn validate<Ctx>(ctx_a: &Ctx, msg: &MsgTimeout) -> Result<(), ContextError>
+pub(crate) enum TimeoutMsgType {
+    Timeout(MsgTimeout),
+    TimeoutOnClose(MsgTimeoutOnClose),
+}
+
+pub(crate) fn timeout_packet_validate<ValCtx>(
+    ctx_a: &ValCtx,
+    module_id: ModuleId,
+    timeout_msg_type: TimeoutMsgType,
+) -> Result<(), ContextError>
+where
+    ValCtx: ValidationContext,
+{
+    match &timeout_msg_type {
+        TimeoutMsgType::Timeout(msg) => validate(ctx_a, msg),
+        TimeoutMsgType::TimeoutOnClose(msg) => timeout_on_close::validate(ctx_a, msg),
+    }?;
+
+    let module = ctx_a
+        .get_route(&module_id)
+        .ok_or(ChannelError::RouteNotFound)?;
+
+    let (packet, signer) = match timeout_msg_type {
+        TimeoutMsgType::Timeout(msg) => (msg.packet, msg.signer),
+        TimeoutMsgType::TimeoutOnClose(msg) => (msg.packet, msg.signer),
+    };
+
+    module
+        .on_timeout_packet_validate(&packet, &signer)
+        .map_err(ContextError::PacketError)
+}
+
+pub(crate) fn timeout_packet_execute<ExecCtx>(
+    ctx_a: &mut ExecCtx,
+    module_id: ModuleId,
+    timeout_msg_type: TimeoutMsgType,
+) -> Result<(), ContextError>
+where
+    ExecCtx: ExecutionContext,
+{
+    let (packet, signer) = match timeout_msg_type {
+        TimeoutMsgType::Timeout(msg) => (msg.packet, msg.signer),
+        TimeoutMsgType::TimeoutOnClose(msg) => (msg.packet, msg.signer),
+    };
+    let chan_end_path_on_a = ChannelEndPath::new(&packet.port_id_on_a, &packet.chan_id_on_a);
+    let chan_end_on_a = ctx_a.channel_end(&chan_end_path_on_a)?;
+
+    // In all cases, this event is emitted
+    let event = IbcEvent::TimeoutPacket(TimeoutPacket::new(packet.clone(), chan_end_on_a.ordering));
+    ctx_a.emit_ibc_event(IbcEvent::Message(MessageEvent::Channel));
+    ctx_a.emit_ibc_event(event);
+
+    let commitment_path_on_a =
+        CommitmentPath::new(&packet.port_id_on_a, &packet.chan_id_on_a, packet.seq_on_a);
+
+    // check if we're in the NO-OP case
+    if ctx_a.get_packet_commitment(&commitment_path_on_a).is_err() {
+        // This error indicates that the timeout has already been relayed
+        // or there is a misconfigured relayer attempting to prove a timeout
+        // for a packet never sent. Core IBC will treat this error as a no-op in order to
+        // prevent an entire relay transaction from failing and consuming unnecessary fees.
+        return Ok(());
+    };
+
+    let module = ctx_a
+        .get_route_mut(&module_id)
+        .ok_or(ChannelError::RouteNotFound)?;
+
+    let (extras, cb_result) = module.on_timeout_packet_execute(&packet, &signer);
+
+    cb_result?;
+
+    // apply state changes
+    let chan_end_on_a = {
+        let commitment_path_on_a = CommitmentPath {
+            port_id: packet.port_id_on_a.clone(),
+            channel_id: packet.chan_id_on_a.clone(),
+            sequence: packet.seq_on_a,
+        };
+        ctx_a.delete_packet_commitment(&commitment_path_on_a)?;
+
+        if let Order::Ordered = chan_end_on_a.ordering {
+            let mut chan_end_on_a = chan_end_on_a;
+            chan_end_on_a.state = State::Closed;
+            ctx_a.store_channel(&chan_end_path_on_a, chan_end_on_a.clone())?;
+
+            chan_end_on_a
+        } else {
+            chan_end_on_a
+        }
+    };
+
+    // emit events and logs
+    {
+        ctx_a.log_message("success: packet timeout".to_string());
+
+        if let Order::Ordered = chan_end_on_a.ordering {
+            let conn_id_on_a = chan_end_on_a.connection_hops()[0].clone();
+
+            let event = IbcEvent::ChannelClosed(ChannelClosed::new(
+                packet.port_id_on_a.clone(),
+                packet.chan_id_on_a.clone(),
+                chan_end_on_a.counterparty().port_id.clone(),
+                chan_end_on_a.counterparty().channel_id.clone(),
+                conn_id_on_a,
+                chan_end_on_a.ordering,
+            ));
+            ctx_a.emit_ibc_event(IbcEvent::Message(MessageEvent::Channel));
+            ctx_a.emit_ibc_event(event);
+        }
+
+        for module_event in extras.events {
+            ctx_a.emit_ibc_event(IbcEvent::Module(module_event));
+        }
+
+        for log_message in extras.log {
+            ctx_a.log_message(log_message);
+        }
+    }
+
+    Ok(())
+}
+
+fn validate<Ctx>(ctx_a: &Ctx, msg: &MsgTimeout) -> Result<(), ContextError>
 where
     Ctx: ValidationContext,
 {
@@ -142,10 +273,7 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::core::ics04_channel::commitment::compute_packet_commitment;
-    use crate::core::timestamp::Timestamp;
-    use crate::core::ExecutionContext;
-    use crate::prelude::*;
+    use super::*;
     use rstest::*;
 
     use crate::core::ics02_client::height::Height;
@@ -160,24 +288,35 @@ mod tests {
     use crate::core::ics04_channel::msgs::timeout::MsgTimeout;
     use crate::core::ics04_channel::Version;
     use crate::core::ics24_host::identifier::{ChannelId, ClientId, ConnectionId, PortId};
+    use crate::core::timestamp::Timestamp;
     use crate::core::timestamp::ZERO_DURATION;
-    use crate::mock::context::MockContext;
 
-    pub struct Fixture {
-        pub context: MockContext,
-        pub client_height: Height,
-        pub msg: MsgTimeout,
-        pub packet_commitment: PacketCommitment,
-        pub conn_end_on_a: ConnectionEnd,
-        pub chan_end_on_a_unordered: ChannelEnd,
-        pub chan_end_on_a_ordered: ChannelEnd,
+    use crate::applications::transfer::MODULE_ID_STR;
+    use crate::mock::context::MockContext;
+    use crate::test_utils::DummyTransferModule;
+
+    struct Fixture {
+        ctx: MockContext,
+        client_height: Height,
+        module_id: ModuleId,
+        msg: MsgTimeout,
+        packet_commitment: PacketCommitment,
+        conn_end_on_a: ConnectionEnd,
+        chan_end_on_a_ordered: ChannelEnd,
+        chan_end_on_a_unordered: ChannelEnd,
     }
 
     #[fixture]
     fn fixture() -> Fixture {
-        let context = MockContext::default();
+        let client_height = Height::new(0, 2).unwrap();
+        let mut ctx = MockContext::default().with_client(&ClientId::default(), client_height);
 
         let client_height = Height::new(0, 2).unwrap();
+
+        let module_id: ModuleId = ModuleId::new(MODULE_ID_STR.to_string());
+        let module = DummyTransferModule::new();
+        ctx.add_route(module_id.clone(), module).unwrap();
+
         let msg_proof_height = 2;
         let msg_timeout_height = 5;
         let timeout_timestamp = Timestamp::now().nanoseconds();
@@ -188,6 +327,7 @@ mod tests {
             timeout_timestamp,
         ))
         .unwrap();
+
         let packet = msg.packet.clone();
 
         let packet_commitment = compute_packet_commitment(
@@ -222,26 +362,27 @@ mod tests {
         .unwrap();
 
         Fixture {
-            context,
+            ctx,
             client_height,
+            module_id,
             msg,
             packet_commitment,
             conn_end_on_a,
-            chan_end_on_a_unordered,
             chan_end_on_a_ordered,
+            chan_end_on_a_unordered,
         }
     }
 
     #[rstest]
     fn timeout_fail_no_channel(fixture: Fixture) {
         let Fixture {
-            context,
+            ctx,
             msg,
             client_height,
             ..
         } = fixture;
-        let context = context.with_client(&ClientId::default(), client_height);
-        let res = validate(&context, &msg);
+        let ctx = ctx.with_client(&ClientId::default(), client_height);
+        let res = validate(&ctx, &msg);
 
         assert!(
             res.is_err(),
@@ -252,7 +393,7 @@ mod tests {
     #[rstest]
     fn timeout_fail_no_consensus_state_for_height(fixture: Fixture) {
         let Fixture {
-            context,
+            ctx,
             msg,
             chan_end_on_a_unordered,
             conn_end_on_a,
@@ -262,7 +403,7 @@ mod tests {
 
         let packet = msg.packet.clone();
 
-        let context = context
+        let ctx = ctx
             .with_channel(
                 PortId::default(),
                 ChannelId::default(),
@@ -276,7 +417,7 @@ mod tests {
                 packet_commitment,
             );
 
-        let res = validate(&context, &msg);
+        let res = validate(&ctx, &msg);
 
         assert!(
             res.is_err(),
@@ -287,7 +428,7 @@ mod tests {
     #[rstest]
     fn timeout_fail_proof_timeout_not_reached(fixture: Fixture) {
         let Fixture {
-            context,
+            ctx,
             mut msg,
             chan_end_on_a_unordered,
             conn_end_on_a,
@@ -307,7 +448,7 @@ mod tests {
 
         let packet = msg.packet.clone();
 
-        let mut context = context
+        let mut ctx = ctx
             .with_client(&ClientId::default(), client_height)
             .with_connection(ConnectionId::default(), conn_end_on_a)
             .with_channel(
@@ -322,22 +463,20 @@ mod tests {
                 packet_commitment,
             );
 
-        context
-            .store_update_time(
-                ClientId::default(),
-                client_height,
-                Timestamp::from_nanoseconds(5).unwrap(),
-            )
-            .unwrap();
-        context
-            .store_update_height(
-                ClientId::default(),
-                client_height,
-                Height::new(0, 4).unwrap(),
-            )
-            .unwrap();
+        ctx.store_update_time(
+            ClientId::default(),
+            client_height,
+            Timestamp::from_nanoseconds(5).unwrap(),
+        )
+        .unwrap();
+        ctx.store_update_height(
+            ClientId::default(),
+            client_height,
+            Height::new(0, 4).unwrap(),
+        )
+        .unwrap();
 
-        let res = validate(&context, &msg);
+        let res = validate(&ctx, &msg);
 
         assert!(
             res.is_err(),
@@ -349,13 +488,13 @@ mod tests {
     #[rstest]
     fn timeout_success_no_packet_commitment(fixture: Fixture) {
         let Fixture {
-            context,
+            ctx,
             msg,
             conn_end_on_a,
             chan_end_on_a_unordered,
             ..
         } = fixture;
-        let context = context
+        let ctx = ctx
             .with_channel(
                 PortId::default(),
                 ChannelId::default(),
@@ -363,7 +502,7 @@ mod tests {
             )
             .with_connection(ConnectionId::default(), conn_end_on_a);
 
-        let res = validate(&context, &msg);
+        let res = validate(&ctx, &msg);
 
         assert!(
             res.is_ok(),
@@ -372,9 +511,9 @@ mod tests {
     }
 
     #[rstest]
-    fn timeout_success_unordered_channel(fixture: Fixture) {
+    fn timeout_unordered_channel_validate(fixture: Fixture) {
         let Fixture {
-            context,
+            ctx,
             msg,
             chan_end_on_a_unordered,
             conn_end_on_a,
@@ -385,7 +524,7 @@ mod tests {
 
         let packet = msg.packet.clone();
 
-        let mut context = context
+        let mut ctx = ctx
             .with_client(&ClientId::default(), client_height)
             .with_connection(ConnectionId::default(), conn_end_on_a)
             .with_channel(
@@ -400,30 +539,28 @@ mod tests {
                 packet_commitment,
             );
 
-        context
-            .store_update_time(
-                ClientId::default(),
-                client_height,
-                Timestamp::from_nanoseconds(1000).unwrap(),
-            )
-            .unwrap();
-        context
-            .store_update_height(
-                ClientId::default(),
-                client_height,
-                Height::new(0, 5).unwrap(),
-            )
-            .unwrap();
+        ctx.store_update_time(
+            ClientId::default(),
+            client_height,
+            Timestamp::from_nanoseconds(1000).unwrap(),
+        )
+        .unwrap();
+        ctx.store_update_height(
+            ClientId::default(),
+            client_height,
+            Height::new(0, 5).unwrap(),
+        )
+        .unwrap();
 
-        let res = validate(&context, &msg);
+        let res = validate(&ctx, &msg);
 
         assert!(res.is_ok(), "Good parameters for unordered channels")
     }
 
     #[rstest]
-    fn timeout_success_ordered_channel(fixture: Fixture) {
+    fn timeout_ordered_channel_validate(fixture: Fixture) {
         let Fixture {
-            context,
+            ctx,
             msg,
             chan_end_on_a_ordered,
             conn_end_on_a,
@@ -434,7 +571,7 @@ mod tests {
 
         let packet = msg.packet.clone();
 
-        let mut context = context
+        let mut ctx = ctx
             .with_client(&ClientId::default(), client_height)
             .with_connection(ConnectionId::default(), conn_end_on_a)
             .with_channel(
@@ -449,23 +586,102 @@ mod tests {
                 packet_commitment,
             );
 
-        context
-            .store_update_time(
-                ClientId::default(),
-                client_height,
-                Timestamp::from_nanoseconds(1000).unwrap(),
-            )
-            .unwrap();
-        context
-            .store_update_height(
-                ClientId::default(),
-                client_height,
-                Height::new(0, 4).unwrap(),
-            )
-            .unwrap();
+        ctx.store_update_time(
+            ClientId::default(),
+            client_height,
+            Timestamp::from_nanoseconds(1000).unwrap(),
+        )
+        .unwrap();
+        ctx.store_update_height(
+            ClientId::default(),
+            client_height,
+            Height::new(0, 4).unwrap(),
+        )
+        .unwrap();
 
-        let res = validate(&context, &msg);
+        let res = validate(&ctx, &msg);
 
         assert!(res.is_ok(), "Good parameters for unordered channels")
+    }
+
+    #[rstest]
+    fn timeout_unordered_chan_execute(fixture: Fixture) {
+        let Fixture {
+            ctx,
+            module_id,
+            msg,
+            packet_commitment,
+            conn_end_on_a,
+            chan_end_on_a_unordered,
+            ..
+        } = fixture;
+        let mut ctx = ctx
+            .with_channel(
+                PortId::default(),
+                ChannelId::default(),
+                chan_end_on_a_unordered,
+            )
+            .with_connection(ConnectionId::default(), conn_end_on_a)
+            .with_packet_commitment(
+                msg.packet.port_id_on_a.clone(),
+                msg.packet.chan_id_on_a.clone(),
+                msg.packet.seq_on_a,
+                packet_commitment,
+            );
+
+        let res = timeout_packet_execute(&mut ctx, module_id, TimeoutMsgType::Timeout(msg));
+
+        assert!(res.is_ok());
+
+        // Unordered channels only emit one event
+        assert_eq!(ctx.events.len(), 2);
+        assert!(matches!(
+            ctx.events[0],
+            IbcEvent::Message(MessageEvent::Channel)
+        ));
+        assert!(matches!(ctx.events[1], IbcEvent::TimeoutPacket(_)));
+    }
+
+    #[rstest]
+    fn timeout_ordered_chan_execute(fixture: Fixture) {
+        let Fixture {
+            ctx,
+            module_id,
+            msg,
+            packet_commitment,
+            conn_end_on_a,
+            chan_end_on_a_ordered,
+            ..
+        } = fixture;
+        let mut ctx = ctx
+            .with_channel(
+                PortId::default(),
+                ChannelId::default(),
+                chan_end_on_a_ordered,
+            )
+            .with_connection(ConnectionId::default(), conn_end_on_a)
+            .with_packet_commitment(
+                msg.packet.port_id_on_a.clone(),
+                msg.packet.chan_id_on_a.clone(),
+                msg.packet.seq_on_a,
+                packet_commitment,
+            );
+
+        let res = timeout_packet_execute(&mut ctx, module_id, TimeoutMsgType::Timeout(msg));
+
+        assert!(res.is_ok());
+
+        // Ordered channels emit 2 events
+        assert_eq!(ctx.events.len(), 4);
+        assert!(matches!(
+            ctx.events[0],
+            IbcEvent::Message(MessageEvent::Channel)
+        ));
+        assert!(matches!(ctx.events[1], IbcEvent::TimeoutPacket(_)));
+        assert!(matches!(
+            ctx.events[2],
+            IbcEvent::Message(MessageEvent::Channel)
+        ));
+        assert!(matches!(ctx.events[3], IbcEvent::ChannelClosed(_)));
     }
 }
